+++ fragments/disabled/mochakeyboard.sh
@@ -1,7 +1,0 @@
-mochakeyboard)
-    name="Mocha Keyboard"
-    type="appInDmgInZip"
-    downloadURL="https://mochasoft.dk/mochakeyboard.dmg.zip"
-    appNewVersion=""
-    expectedTeamID="RR9F5EPNVW"
-    ;;