+++ fragments/disabled/mochatn3270.sh
@@ -1,7 +1,0 @@
-mochatn3270)
-    name="TN3270"
-    type="appInDmgInZip"
-    downloadURL="https://mochasoft.dk/tn3270.dmg.zip"
-    appNewVersion=""
-    expectedTeamID="RR9F5EPNVW"
-    ;;