+++ fragments/labels/mochatn3270.sh
@@ -1,0 +1,16 @@
+mochatn3270)
+    name="TN3270"
+    type="appInDmgInZip"
+    downloadURL="https://mochasoft.dk/tn3270.dmg.zip"
+    curlOptions=( -H "User-Agent: Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/15.1 Safari/605.1.15"
+-H "accept: text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9"
+-H "accept-encoding: gzip, deflate, br"
+-H "accept-language: en-US,en;q=0.9"
+-H "sec-fetch-dest: document"
+-H "sec-fetch-mode: navigate"
+-H "sec-fetch-user: ?1"
+-H "sec-gpc: 1"
+-H "upgrade-insecure-requests: 1" )
+    appNewVersion=""
+    expectedTeamID="RR9F5EPNVW"
+    ;;